requirements.txt
langchain
langchain-groq
python-dotenv
langchain_community
langchain-mongodb
langchain_openai
pymongo
jq
